requirements.txt
openai>=1.5.0
python-dotenv>=0.19.0
pillow>=8.0.0
